+++ src/main/java/com/iip/ui/feature_extraction/controller/FE1ParticipleController.java
@@ -1,10 +1,8 @@
 package com.iip.ui.feature_extraction.controller;
 
 import com.iip.ui.feature_extraction.Main;
-import com.iip.ui.feature_extraction.execute.Doc2vec;
 import com.iip.ui.feature_extraction.execute.Participle;
 import com.iip.ui.feature_extraction.execute.connection.DatabaseOperations;
-import com.iip.ui.feature_extraction.execute.textrank.TextRank;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -15,7 +13,7 @@ import javafx.scene.input.MouseEvent;
 import java.util.ArrayList;
 import java.util.List;
 
-public class FEStep3Controller {
+public class FE1ParticipleController {
 
     @FXML
     private ListView<String> rawDataListView;
@@ -26,9 +24,9 @@ public class FEStep3Controller {
     private ObservableList<String> handledDataList = FXCollections.observableArrayList();
 
     @FXML
-    private Label dbname, dbname_, dataOrigin, dataOutput, hint;
+    private Label dbname, dbname_, dataOrigin, dataOutput;
 
-    public static List<String> docs = new ArrayList<String>();
+    private List<String> docs = new ArrayList<String>();
     private List<String> partedDocs = new ArrayList<String>();
 
     public void initialize(){
@@ -48,7 +46,7 @@ public class FEStep3Controller {
     @FXML
     private void handleDataClicked(MouseEvent event){
         if(partedDocs.size()>0) return;
-        handledDataList.addAll(getKeyedDocs(docs));
+        handledDataList.addAll(getPartedDocs(docs));
         handledDataListView.setItems(handledDataList);
     }
 
@@ -59,7 +57,7 @@ public class FEStep3Controller {
             Main.f_alert_informationDialog("操作失败", "数据库还没链接好!");
             return;
         }
-        DatabaseOperations.write(DatabaseOperations.outputTableNames[2], DatabaseOperations.dataColumn, partedDocs);
+        DatabaseOperations.write(DatabaseOperations.outputTableNames[0], DatabaseOperations.dataColumn, partedDocs);
         DatabaseOperations.print("导出表完成!");
         Main.f_alert_informationDialog("操作成功", "导出表完成!");
     }
@@ -73,21 +71,19 @@ public class FEStep3Controller {
             Main.f_alert_informationDialog("操作失败", "数据库还没链接好!");
             return docs;
         }
-        docs = DatabaseOperations.read(DatabaseOperations.originalDataTables[2], DatabaseOperations.dataColumn);
+        docs = DatabaseOperations.read(DatabaseOperations.originalDataTables[0], DatabaseOperations.dataColumn);
         dbname.setText("数据库: "+DatabaseOperations.DBNAME);
         dbname_.setText("数据库: "+DatabaseOperations.DBNAME);
-        dataOrigin.setText("源数据表格: "+DatabaseOperations.originalDataTables[2]);
-        dataOutput.setText("导出数据表格: "+DatabaseOperations.outputTableNames[2]);
+        dataOrigin.setText("源数据表格: "+DatabaseOperations.originalDataTables[0]);
+        dataOutput.setText("导出数据表格: "+DatabaseOperations.outputTableNames[0]);
         DatabaseOperations.print("导入表完成!");
         Main.f_alert_informationDialog("操作成功", "导入表完成!");
         return docs;
     }
 
-    //选取关键词操作
-    private List<String> getKeyedDocs(List<String> docs){
-        hint.setText("该操作比较耗时，请稍等......");
-        partedDocs = Doc2vec.getVecDocs(docs, DatabaseOperations.path);
-        hint.setText("操作完成!");
+    //分词操作
+    private List<String> getPartedDocs(List<String> docs){
+        partedDocs = Participle.getPartedDocs(docs);
         return partedDocs;
     }
